+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Removed `selectionIndicator` property from `LYRUIParticipantTableViewController`.
 * Added `participantPickerController:didDeselectParticipant:` to `LYRUIParticipantPickerControllerDelegate`.
 * Added `participantTableViewController:didDeselectParticipant:` to `LYRUIParticipantTableViewControllerDelegate`.
+* Replaced `presentParticipant:`, `updateWithSortType:` and `shouldShowAvatarImage:` on `LYRUIParticipantPresenting` with `presentParticipant:withSortType:shouldShowAvatarImage:`.
 
 ### Bug Fixes